test(sase): split orchestration API tests into per-resource modules (<1000 lines)
Reorganize test_sase_api_orchestration.py (1037 lines) into three cohesive
per-resource modules for maintainability:

- test_sase_api_clusters.py (481 lines): cluster registration, heartbeat,
  headend-config, per-cluster credential auth, tenant-mismatch, >5-node gate
- test_sase_api_clients.py (371 lines): client registration, config,
  tunnel-config, key rotation, client + headend metrics, list, isolation
- test_sase_api_status.py (107 lines): status endpoint

Move shared fixtures (app_with_sase, tokens, bootstrap) to conftest.py
for reuse. All 21 tests pass; full suite remains at 388 tests.
